Major improvements to error handling and reporting.
Fixed a bug where the translator would claim that the string had disappeared if it was actually
just unverified. I think this got introduced ages ago when Mark added the verify feature.

Instead, now it figures out the "problem" with the translation and whether it should be an error
or just an info log. The id and problem are printed to the appropriate log level.

"Problems" include being untranslated, being unverified, and having problems with braces. Braces
are checked for bad formatting, missing parameters that are present in the english, and extra
parameters not present in the english.

Also, _errors, _placeholders and _retained counts are accumulated. This will be a big help to
developers since if these counters are non-zero, then translations are not complete.